requirements.txt
ultralytics
pyyaml
opencv-python
beautifulsoup4
huggingface-hub
langchain_community 
python-dotenv
fastapi
uvicorn
gtts
transformers
